+++ ichnaea/service/geosubmit/views.py
@@ -4,7 +4,6 @@ from pyramid.httpexceptions import HTTPNotFound, HTTPOk
 from pytz import utc
 
 from ichnaea.customjson import dumps
-from ichnaea.geocalc import location_is_in_country
 from ichnaea.service.base import check_api_key
 from ichnaea.service.error import (
     JSONParseError,
@@ -125,26 +124,12 @@ def configure_geosubmit(config):
     config.add_view(geosubmit_view, route_name='v1_geosubmit', renderer='json')
 
 
-def check_geoip(request, data):
+def flatten_items(data):
     if any(data.get('items', ())):
         items = data['items']
     else:
         items = [data]
 
-    # Verify that the submission comes from the same country as the lat/lon.
-    if request.client_addr:
-        geoip = request.registry.geoip_db.geoip_lookup(request.client_addr)
-        if geoip:
-            filtered_items = []
-            for item in items:
-                lat = float(item['latitude'])
-                lon = float(item['longitude'])
-                country = geoip['country_code']
-                if location_is_in_country(lat, lon, country):
-                    filtered_items.append(item)
-                else:
-                    get_stats_client().incr("submit.geoip_mismatch")
-            return filtered_items
     return items
 
 
@@ -168,7 +153,7 @@ def geosubmit_view(request):
 
 def process_batch(request, data, errors):
     nickname = request.headers.get('X-Nickname', u'')
-    upload_items = check_geoip(request, data)
+    upload_items = flatten_items(data)
     validated, errors = process_upload(nickname, upload_items)
 
     if errors:
@@ -200,7 +185,7 @@ def process_single(request):
     data = {'items': [data]}
 
     nickname = request.headers.get('X-Nickname', u'')
-    upload_items = check_geoip(request, data)
+    upload_items = flatten_items(data)
     validated, errors = process_upload(nickname, upload_items)
 
     if errors: